Move memoized to utils
