+++ lib/markdown2.py
@@ -112,7 +112,15 @@ import codecs
 from collections import defaultdict
 
 from lib.errors import MarkdownError
-from lib.utils import slugify, calculate_toc_html, curry, regex_from_encoded_pattern, dedentlines, dedent
+from lib.utils import (
+    slugify,
+    calculate_toc_html,
+    curry,
+    regex_from_encoded_pattern,
+    dedentlines,
+    dedent,
+    memoized,
+)
 
 # ---- globals
 
@@ -2471,33 +2479,6 @@ class UnicodeWithAttrs(str):
     toc_html = None
 
 
-class _memoized(object):
-    """Decorator that caches a function's return value each time it is called.
-    If called later with the same arguments, the cached value is returned, and
-    not re-evaluated.
-
-    http://wiki.python.org/moin/PythonDecoratorLibrary
-    """
-    def __init__(self, func):
-        self.func = func
-        self.cache = {}
-
-    def __call__(self, *args):
-        try:
-            return self.cache[args]
-        except KeyError:
-            self.cache[args] = value = self.func(*args)
-            return value
-        except TypeError:
-            # uncachable -- for instance, passing a list as an argument.
-            # Better to not cache than to blow up entirely.
-            return self.func(*args)
-
-    def __repr__(self):
-        """Return the function's docstring."""
-        return self.func.__doc__
-
-
 def _xml_oneliner_re_from_tab_width(tab_width):
     """Standalone XML processing instruction regex."""
     return re.compile(r"""
@@ -2517,7 +2498,7 @@ def _xml_oneliner_re_from_tab_width(tab_width):
             (?=\n{2,}|\Z)       # followed by a blank line or end of document
         )
         """ % (tab_width - 1), re.X)
-_xml_oneliner_re_from_tab_width = _memoized(_xml_oneliner_re_from_tab_width)
+_xml_oneliner_re_from_tab_width = memoized(_xml_oneliner_re_from_tab_width)
 
 
 def _hr_tag_re_from_tab_width(tab_width):
@@ -2537,7 +2518,7 @@ def _hr_tag_re_from_tab_width(tab_width):
             (?=\n{2,}|\Z)       # followed by a blank line or end of document
         )
         """ % (tab_width - 1), re.X)
-_hr_tag_re_from_tab_width = _memoized(_hr_tag_re_from_tab_width)
+_hr_tag_re_from_tab_width = memoized(_hr_tag_re_from_tab_width)
 
 
 def _xml_escape_attr(attr, skip_single_quote=True):
